% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_blocks}
\alias{census_blocks}
\title{Census Blocks for Wake County, North Carolina}
\format{
An sf object with variables:
\describe{
  \item{GEOID}{Character. The unique identifier for the census block}
  \item{geometry}{sfc_MULTIPOLYGON. The geometry of the census block boundary}
}
}
\source{
Walker K, Herman M (2024). \emph{tidycensus: Load US Census Boundary and Attribute
  Data as 'tidyverse' and 'sf'-Ready Data Frames}. R package version 1.6.5,
  \url{https://walker-data.com/tidycensus/}.
}
\usage{
census_blocks
}
\description{
A dataset containing the boundaries of census blocks in Wake County, North Carolina.
}
\keyword{datasets}
