% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_msas}
\alias{census_msas}
\title{United States Metropolitan Statistical Areas (MSAs)}
\format{
An sf object with variables:
\describe{
  \item{GEOID}{Character. The unique identifier for the MSA}
  \item{NAME}{Character. The name of the MSA}
  \item{geometry}{sfc_MULTIPOLYGON. The geometry of the MSA boundary}
}
}
\source{
Walker K, Herman M (2024). \emph{tidycensus: Load US Census Boundary and Attribute
  Data as 'tidyverse' and 'sf'-Ready Data Frames}. R package version 1.6.5,
  \url{https://walker-data.com/tidycensus/}.
}
\usage{
census_msas
}
\description{
A dataset containing the boundaries of Metropolitan Statistical Areas in the United States.
}
\keyword{datasets}
