% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_tracts}
\alias{census_tracts}
\title{Census Tracts for North Carolina}
\format{
An sf object with variables:
\describe{
  \item{GEOID}{Character. The unique identifier for the census tract}
  \item{NAME}{Character. The name or number of the census tract}
  \item{geometry}{sfc_MULTIPOLYGON. The geometry of the census tract boundary}
}
}
\source{
Walker K, Herman M (2024). \emph{tidycensus: Load US Census Boundary and Attribute
  Data as 'tidyverse' and 'sf'-Ready Data Frames}. R package version 1.6.5,
  \url{https://walker-data.com/tidycensus/}.
}
\usage{
census_tracts
}
\description{
A dataset containing the boundaries of census tracts in North Carolina.
}
\keyword{datasets}
