% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dews_boundaries.R
\name{create_dews_boundaries}
\alias{create_dews_boundaries}
\title{Create NIDIS DEWS Boundaries}
\usage{
create_dews_boundaries(
  us_states,
  missouri_basin,
  pnw_basins,
  dews_regions,
  extended = FALSE
)
}
\arguments{
\item{us_states}{sf object of US states from tidycensus}

\item{missouri_basin}{sf object of Missouri River Basin from USGS WBD}

\item{pnw_basins}{sf object of Pacific Northwest river basins from USGS WBD}

\item{dews_regions}{list of DEWS regions and their component states}

\item{extended}{logical. If TRUE, includes extended regions (marked with *). Default FALSE.}
}
\value{
sf object containing DEWS boundaries with names and geometries
}
\description{
This function creates an sf object of NIDIS Drought Early Warning System (DEWS) boundaries
by combining state boundaries and river basin components where necessary.
}
