% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nclimgrid_data.R
\name{get_nclimgrid_data}
\alias{get_nclimgrid_data}
\title{Retrieve nClimGrid-Daily Data}
\usage{
get_nclimgrid_data(
  start_date,
  end_date,
  variables,
  boundary = NULL,
  boundary_type = NULL,
  boundary_name = NULL,
  use_scaled = TRUE
)
}
\arguments{
\item{start_date}{Character string. Start date in "YYYY-MM-DD" format. Must be 1951-01-01 or later.}

\item{end_date}{Character string. End date in "YYYY-MM-DD" format.}

\item{variables}{Character vector. Any combination of "tmax", "tmin", "tavg", or "prcp".}

\item{boundary}{sf object. A spatial boundary to filter the data. If NULL, a boundary_type must be specified.}

\item{boundary_type}{Character string. One of "custom", "state", "county", "msa", "tract", "block", or "dews".}

\item{boundary_name}{Character string. Name of the boundary (e.g., state name) if using a predefined boundary type.}

\item{use_scaled}{Logical. If TRUE, use scaled data. If FALSE, use preliminary data. Default is TRUE.}
}
\value{
A stars object containing the filtered nClimGrid-Daily data.
}
\description{
This function retrieves nClimGrid-Daily data from the AWS S3 bucket, allowing for
temporal and spatial filtering using various boundary types. It is recommended to use
this function for retrieving data over longer time periods or larger spatial areas,
rather than for single-day or single-point values.
}
\examples{
\dontrun{
# Using a predefined state boundary for a month
data <- get_nclimgrid_data("2020-01-01", "2020-01-31", c("tmax", "tmin"),
                           boundary_type = "state", boundary_name = "North Carolina")
}
}
