% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nidis_dews}
\alias{nidis_dews}
\title{NIDIS Drought Early Warning System (DEWS) Boundaries}
\format{
An sf object with 2 variables:
\describe{
  \item{DEWS}{character. Name of the DEWS region}
  \item{geometry}{sfc_MULTIPOLYGON. Spatial geometry of the DEWS region}
}
}
\source{
State boundaries from US Census Bureau via tidycensus package
  \url{https://walker-data.com/tidycensus/}

  River basin boundaries from USGS Watershed Boundary Dataset (WBD)
  \url{https://www.usgs.gov/national-hydrography/watershed-boundary-dataset}

  DEWS region definitions from NIDIS
  \url{https://www.drought.gov/dews}
}
\usage{
nidis_dews
}
\description{
A spatial dataset containing the official boundaries of the National Integrated
Drought Information System (NIDIS) Drought Early Warning System (DEWS) regions.
The boundaries are constructed from a combination of state boundaries and river basins.
}
\details{
The DEWS regions include:
  \itemize{
    \item California-Nevada (state-based)
    \item Midwest (state-based)
    \item Northeast (state-based)
    \item Pacific Northwest (states + Upper Columbia and Kootenai-Pend Oreille-Spokane basins)
    \item Southeast (state-based)
    \item Southern Plains (state-based)
    \item Intermountain West (state-based)
    \item Missouri River Basin (Missouri River Basin + ND and SD)
  }
}
\keyword{datasets}
