% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nidis_dews_ext}
\alias{nidis_dews_ext}
\title{Extended NIDIS Drought Early Warning System (DEWS) Boundaries}
\format{
An sf object with 2 variables:
\describe{
  \item{DEWS}{character. Name of the DEWS region}
  \item{geometry}{sfc_MULTIPOLYGON. Spatial geometry of the DEWS region}
}
}
\source{
State boundaries from US Census Bureau via tidycensus package
  \url{https://walker-data.com/tidycensus/}

  River basin boundaries from USGS Watershed Boundary Dataset (WBD)
  \url{https://www.usgs.gov/national-hydrography/watershed-boundary-dataset}

  DEWS region definitions from NIDIS
  \url{https://www.drought.gov/dews}
}
\usage{
nidis_dews_ext
}
\description{
A spatial dataset containing both official and extended boundaries of the National
Integrated Drought Information System (NIDIS) Drought Early Warning System (DEWS)
regions. The extended boundaries include additional regions to provide complete
coverage of the contiguous United States.
}
\details{
Includes all regions from \code{nidis_dews} plus:
  \itemize{
    \item Mid Atlantic (WV, PA, MD, NJ)
    \item Lower Mississippi (LA, AK, MS)
  }
}
\keyword{datasets}
