% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_counties}
\alias{us_counties}
\title{United States County Boundaries}
\format{
An sf object with approximately 3,142 rows and 4 variables:
\describe{
  \item{NAME}{Character. The name of the county}
  \item{STATE}{Character. The two-letter abbreviation of the state in which the county is located}
  \item{GEOID}{Character. The unique identifier for the county}
  \item{geometry}{sfc_MULTIPOLYGON. The geometry of the county boundary}
}
}
\source{
Walker K, Herman M (2024). \emph{tidycensus: Load US Census Boundary and Attribute
  Data as 'tidyverse' and 'sf'-Ready Data Frames}. R package version 1.6.5,
  \url{https://walker-data.com/tidycensus/}.
}
\usage{
us_counties
}
\description{
A dataset containing the boundaries of all counties in the United States.
}
\keyword{datasets}
