% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{United States State Boundaries}
\format{
An sf object with 51 rows and 3 variables:
\describe{
  \item{STATE}{Character. The two-letter state abbreviation}
  \item{NAME}{Character. The full name of the state}
  \item{geometry}{sfc_MULTIPOLYGON. The geometry of the state boundary}
}
}
\source{
Walker K, Herman M (2024). \emph{tidycensus: Load US Census Boundary and Attribute
  Data as 'tidyverse' and 'sf'-Ready Data Frames}. R package version 1.6.5,
  \url{https://walker-data.com/tidycensus/}.
}
\usage{
us_states
}
\description{
A dataset containing the boundaries of all states in the United States.
}
\keyword{datasets}
